+++ src/CertificateManager/CertificateUtility.cs
@@ -61,12 +61,13 @@ namespace CertificateManager
             {
                 throw new ArgumentException($"{nameof(distinguishedName.CommonName)} must be a valid name", nameof(distinguishedName.CommonName));
             }
-            if (string.IsNullOrEmpty(distinguishedName.Country))
-            {
-                throw new ArgumentException($"{nameof(distinguishedName.Country)} must be a valid country", nameof(distinguishedName.Country));
-            }
 
-            var sb = new StringBuilder($"CN={distinguishedName.CommonName}, C={distinguishedName.Country}");
+            var sb = new StringBuilder($"CN={distinguishedName.CommonName}");
+
+            if (!string.IsNullOrEmpty(distinguishedName.Country))
+            {
+                sb.Append($", C={distinguishedName.Country}");
+            }
 
             if (!string.IsNullOrEmpty(distinguishedName.Organisation))
             {