+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Certificate Manager change log
 
-<a name="2020-01-18"></a>
-## 2020-01-18 version 1.0.0
-* Initial Version
+<a name="2020-01-22"></a>
+## 2020-01-22 version 1.0.0
+* creates client server certificate authentication certificates
+* support for chained certificates: root, intermediate, client, server, device, leaf 
+* support for self signed certificates
+* Device certificates for Azure IoT Hub with chain: .cer .pem .pfx exports
+* Exports .cer .pem .pfx files